Refactor Hotkey and HotkeyManager to use Hotkey object as map key
- Introduced HotkeyModifierEnum and HotkeyModifiers class using Flags.h.
- Simplified HotkeyEnum to only contain base keys.
- Updated Hotkey class to store base key and modifiers separately.
- Implemented std::hash for Hotkey using numeric_hash from hash.h.
- Updated HotkeyManager to use std::unordered_map<Hotkey, std::string>.
- Added keypad aliases in qtKeyToHotkeyBase to fix pre-existing test failures.
- Updated InputWidget and AbstractParser-Hotkey to use new Hotkey API.
- Fixed integer precision warning in std::hash<Hotkey>.
- Preserved existing serialization format for backward compatibility.
- Verified all changes with TestHotkeyManager and applied clang-format.

+++ src/client/Hotkey.h
@@ -250,8 +250,8 @@ struct hash<Hotkey>
 {
     size_t operator()(const Hotkey &hk) const noexcept
     {
-        uint16_t packed = (static_cast<uint16_t>(hk.base()) << 8)
-                          | static_cast<uint8_t>(hk.modifiers());
+        uint16_t packed = static_cast<uint16_t>((static_cast<uint16_t>(hk.base()) << 8)
+                                                | static_cast<uint8_t>(hk.modifiers()));
         return numeric_hash(packed);
     }
 };
